Port to C++ and ReportEvent.
Using clr functionality in node.js bindings leads to crashes sometimes.
Therefore, I ported the library to C++ and using the `ReportEvent`
function of the WIN32 API instead. At the same time, I switched to Nan to
ensure the library is compatible with as many node.js/v8 versions as
possible.