Introduce async log function.
ATTENTION: this is a breaking API change!

The EventLog.log() function is now async and requires a callback function
as it's third argument. The previous sync function is now located at
EventLog.logSync().

Note: The Windows Event Log Handle is thread-safe and can therefore be
used from multiple EventLogAsync workers and the main thread at the same
time.